refactor(<ExplorePage/>): add useExplore
